% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StocksFuturePrice.R
\name{stocks_future_price}
\alias{stocks_future_price}
\title{Probability of Stock's future price}
\usage{
stocks_future_price(GOOG)
}
\arguments{
\item{'GOOG'}{}
}
\description{
The package help to predict stock's future price.
}
\examples{
stocks_future_price('GOOG')

}
